pipeline/requirements.txt
google-cloud-bigquery>=3.0,<4.0
pyarrow>=14.0,<18.0
boto3>=1.34,<2.0
